feat: добавил enum с апи респосами свагер
