+++ meetups/Task_6.md
@@ -12,6 +12,7 @@
 2. Описание;
 3. Набор тегов / ключевых слов;
 4. Время и место проведения.
+5. записаться на митап
 
 ## Стек
 
@@ -23,7 +24,7 @@
 
 1. Валидация DTO (достаточно будет использовать joi).
 2. Переработать запрос на получение списка митапов так, чтобы с его помощью можно было осуществить поиск по митапам, отфильтровать их, отсортировать. Результат также должен быть разбит на страницы. (частично сделал, осталось добавит ьвыбор поля фильтрации и значения поля фильтрации)
-3. Добавить документацию к API через Swagger. (первая версия есть)
+3. Добавить документацию к API через Swagger.
 4. Система аутентификации (предлагается к использованию библиотека PassportJS). Систему аутентификации можно разрабатывать поэтапно, постепенно расширяя и усложняя её:
    - Пользователь регистрируется, авторизуется (получая JWT Access Token). Используя Access Token можно получить информацию о владельце токена (т.е. получить текущего пользователя по токену) и записаться на митап.
    - Добавить персистентный JWT Refresh Token.
