+++ meetups/src/app.module.ts
@@ -7,10 +7,12 @@ import { MeetupModule } from './meetup/meetup.module';
 import { PrismaModule } from './prisma/prisma.module';
 import { TagsModule } from './tags/tags.module';
 import { UsersModule } from './users/users.module';
+import { UsersService } from './users/users.service';
+import { PasswordService } from './auth/password.service';
 
 @Module({
   imports: [UsersModule, MeetupModule, AuthModule, PrismaModule, TagsModule],
   controllers: [AppController],
-  providers: [AppService],
+  providers: [AppService, UsersService, PasswordService],
 })
 export class AppModule {}
