+++ meetups/src/consts/swagger.ts
@@ -1,0 +1,6 @@
+export const EApiResponses = {
+  SUCCESS: { status: 200, description: 'Успешное выполнение' },
+  NOT_AUTH: { status: 401, description: 'Не авторизован' },
+  NOT_PERMISSION: { status: 403, description: 'Нет прав' },
+  SERVER_ERROR: { status: 500, description: 'Ошибка' },
+};
