+++ meetups/src/meetup/meetup.controller.ts
@@ -25,6 +25,7 @@ import { CreateMeetupDto } from './dto/create-meetup.dto';
 import { PaginationDto } from './dto/pagination.dto';
 import { UpdateMeetupDto } from './dto/update-meetup.dto';
 import { MeetupService } from './meetup.service';
+import { EApiResponses } from 'src/consts/swagger';
 
 @ApiTags('meetup')
 @Controller('meetup')
@@ -33,9 +34,10 @@ export class MeetupController {
 
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Создание митапа' })
-  @ApiResponse({ status: 200, description: 'Успешное создание' })
-  @ApiResponse({ status: 401, description: 'Не авторизован' })
-  @ApiResponse({ status: 500, description: 'Ошибка' })
+  @ApiResponse(EApiResponses.SUCCESS)
+  @ApiResponse(EApiResponses.NOT_AUTH)
+  @ApiResponse(EApiResponses.NOT_PERMISSION)
+  @ApiResponse(EApiResponses.SERVER_ERROR)
   @UseGuards(JwtAuthGuard, RolesGuard)
   @Post()
   @Roles(Role.creator)
@@ -50,16 +52,16 @@ export class MeetupController {
   }
 
   @ApiOperation({ summary: 'Получение митапов' })
-  @ApiResponse({ status: 200, description: 'Успешное получение митапов' })
-  @ApiResponse({ status: 500, description: 'Ошибка' })
+  @ApiResponse(EApiResponses.SUCCESS)
+  @ApiResponse(EApiResponses.SERVER_ERROR)
   @Get()
   findAll(@Query() query: PaginationDto) {
     return this.meetupService.findAll(query);
   }
 
   @ApiOperation({ summary: 'Получение митапа по id' })
-  @ApiResponse({ status: 200, description: 'Успешное получение митапа' })
-  @ApiResponse({ status: 500, description: 'Ошибка' })
+  @ApiResponse(EApiResponses.SUCCESS)
+  @ApiResponse(EApiResponses.SERVER_ERROR)
   @Get(':id')
   findOne(@Param('id') id: string) {
     return this.meetupService.findOne(+id);
@@ -67,9 +69,10 @@ export class MeetupController {
 
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Обновление митапа' })
-  @ApiResponse({ status: 200, description: 'Успешное обновление' })
-  @ApiResponse({ status: 401, description: 'Не авторизован' })
-  @ApiResponse({ status: 500, description: 'Ошибка' })
+  @ApiResponse(EApiResponses.SUCCESS)
+  @ApiResponse(EApiResponses.NOT_AUTH)
+  @ApiResponse(EApiResponses.NOT_PERMISSION)
+  @ApiResponse(EApiResponses.SERVER_ERROR)
   @UseGuards(JwtAuthGuard, RolesGuard)
   @Roles(Role.creator)
   @Patch(':id')
@@ -93,9 +96,10 @@ export class MeetupController {
 
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Удаление митапа' })
-  @ApiResponse({ status: 200, description: 'Успешное удаление' })
-  @ApiResponse({ status: 401, description: 'Не авторизован' })
-  @ApiResponse({ status: 500, description: 'Ошибка' })
+  @ApiResponse(EApiResponses.SUCCESS)
+  @ApiResponse(EApiResponses.NOT_AUTH)
+  @ApiResponse(EApiResponses.NOT_PERMISSION)
+  @ApiResponse(EApiResponses.SERVER_ERROR)
   @UseGuards(JwtAuthGuard, RolesGuard)
   @Roles(Role.creator)
   @Delete(':id')
