+++ meetups/src/users/users.controller.ts
@@ -13,37 +13,18 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { JwtAuthGuard } from 'src/auth/guards/jwt/jwt-auth.guard';
-import { PasswordService } from 'src/auth/password.service';
-
-import { CreateUserDto } from './dto/create-user.dto';
-import { UsersService } from './users.service';
+import { EApiResponses } from 'src/consts/swagger';
 @ApiTags('users')
 @Controller('users')
 export class UsersController {
-  constructor(
-    private readonly usersService: UsersService,
-    private readonly passwordService: PasswordService,
-  ) {}
+  constructor() {}
 
-  @ApiOperation({ summary: 'Создание пользователя' })
-  @ApiResponse({ status: 200, description: 'Успешное обновление' })
-  @ApiResponse({ status: 500, description: 'Ошибка' })
-  @Post()
-  async create(@Body() createUserDto: CreateUserDto) {
-    const hashedPassword = await this.passwordService.hashPassword(
-      createUserDto.password,
-    );
-    return this.usersService.create({
-      ...createUserDto,
-      password: hashedPassword,
-    });
-  }
   //TODO вынести в enum статусы ответов
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Просмотр профиля' })
-  @ApiResponse({ status: 200, description: 'Успешное получение провиля' })
-  @ApiResponse({ status: 401, description: 'Не авторизован' })
-  @ApiResponse({ status: 500, description: 'Ошибка' })
+  @ApiResponse(EApiResponses.SUCCESS)
+  @ApiResponse(EApiResponses.NOT_AUTH)
+  @ApiResponse(EApiResponses.SERVER_ERROR)
   @UseGuards(JwtAuthGuard)
   @Get('/profile')
   getProfile(@Request() req) {
